# authentication/config/openapi.yml
openapi: 3.1.0
info:
  title: Authentication API
  version: 1.0.0
  description: API to handle authentication including local and social logins.
servers:
  - url: http://localhost:3003/
    description: Local development server
paths:
  /auth/login:
    post:
      summary: Authenticate a user with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Authentication failed
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '409':
          description: Username or email already exists
...
  /auth/mobile/login:
    post:
      summary: Create a bearer-token session for a native mobile client
      description: Returns tokens in the JSON body and never sets authentication cookies.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, device_id]
              properties:
                username:
                  type: string
                  description: PokeGoNexus username or email address
                password:
                  type: string
                  format: password
                device_id:
                  type: string
      responses:
        '200':
          description: Mobile session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileSessionResponse'
        '401':
          description: Invalid credentials
...
  /auth/google/complete-registration:
    post:
      summary: Create an account from a verified pending Google identity
      responses:
        '201':
          description: Google account created and session cookies issued
        '409':
          description: Username, email, trainer name, or trainer code already exists
components:
  securitySchemes:
    bearerAuth:    # Assuming JWT Bearer Token is used
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginResponse:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        token:
          type: string
        message:
          type: string
    RegisterResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            username:
              type: string
            email:
              type: string
        token:
          type: string
        message:
          type: string
    MobileSessionResponse:
      type: object
      required:
        - user
        - accessToken
        - refreshToken
        - accessTokenExpiry
        - refreshTokenExpiry
      properties:
        user:
          type: object
          required: [user_id, username, email]
          properties:
            user_id:
              type: string
            username:
              type: string
            email:
              type: string
              format: email
            pokemonGoName:
              type: [string, 'null']
            trainerCode:
              type: [string, 'null']
            allowLocation:
              type: boolean
            location:
              type: [string, 'null']
            coordinates:
              type: [object, 'null']
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiry:
          type: string
          format: date-time
        refreshTokenExpiry:
          type: string
          format: date-time
...